LayoutTests/svg/dom/SVGTests.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("This test checks the SVGTests API");

var svgElement = document.createElementNS("http://www.w3.org/2000/svg", "svg");
var foreignObject = document.createElementNS("http://www.w3.org/2000/svg", "foreignObject");

debug("");
debug("Check the requiredFeatures, requiredExtensions and systemLanguage attributes");
shouldBeTrue("foreignObject.requiredFeatures instanceof SVGStringList");
shouldBeTrue("foreignObject.requiredExtensions instanceof SVGStringList");
shouldBeTrue("foreignObject.systemLanguage instanceof SVGStringList");
successfullyParsed = true;
</script>
</body>
</html>
