LayoutTests/svg/dom/SVGTests-expected.txt

This test checks the SVGTests API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check the requiredFeatures, requiredExtensions and systemLanguage attributes
PASS foreignObject.requiredFeatures instanceof SVGStringList is true
PASS foreignObject.requiredExtensions instanceof SVGStringList is true
PASS foreignObject.systemLanguage instanceof SVGStringList is true
PASS successfullyParsed is true

TEST COMPLETE
